# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIGORA - Plataforma de Gestão Jurídica
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tarefas:
    get:
      operationId: listTarefas
      tags: [tarefas]
      summary: List tasks
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: processoId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TarefaWithDetails"
    post:
      operationId: createTarefa
      tags: [tarefas]
      summary: Create task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTarefaRequest"
      responses:
        "201":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarefa"
  /tarefas/{id}:
    get:
      operationId: getTarefa
      tags: [tarefas]
      summary: Get task by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Task data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TarefaWithDetails"
        "404":
          description: Task not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateTarefa
      tags: [tarefas]
      summary: Update task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTarefaRequest"
      responses:
        "200":
          description: Task updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tarefa"
    delete:
      operationId: deleteTarefa
      tags: [tarefas]
      summary: Delete task
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Task deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    ProcessoSummary:
      type: object
      properties:
        id:
          type: integer
        numeroProcesso:
          type: string
        area:
          type: string
          nullable: true
        status:
          type: string
      required:
        - id
        - numeroProcesso
        - status
...
    Tarefa:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descricao:
          type: string
          nullable: true
        processoId:
          type: integer
          nullable: true
        prazoId:
          type: integer
          nullable: true
        responsavel:
          type: string
          nullable: true
        dataLimite:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [pendente, em_andamento, concluida, atrasada]
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - titulo
        - status
        - criadoEm
    TarefaWithDetails:
      allOf:
        - $ref: "#/components/schemas/Tarefa"
        - type: object
          properties:
            processo:
              $ref: "#/components/schemas/ProcessoSummary"
              nullable: true
    CreateTarefaRequest:
      type: object
      properties:
        titulo:
          type: string
        descricao:
          type: string
          nullable: true
        processoId:
          type: integer
          nullable: true
        prazoId:
          type: integer
          nullable: true
        responsavel:
          type: string
          nullable: true
        dataLimite:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [pendente, em_andamento, concluida, atrasada]
      required:
        - titulo
        - status
...